requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
dj-database-url==0.5.0
Django==2.1.5
django-bootstrap==0.2.4
django-bootstrap4==0.0.7
flask==0.12.2
gunicorn==19.9.0
idna==2.6
itsdangerous==0.24
jinja2==2.10
markupsafe==1.0
pytelegrambotapi==3.5.2
pytz==2018.9
requests==2.18.4
six==1.11.0
urllib3==1.22
werkzeug==0.14.1
whitenoise==4.1.2
psycopg2==2.7.7
Pillow==6.0.0
django-crispy-forms==1.7.2
django-infinite-pagination==1.0